Moved Package extension to PackageBuilder and sank it into PackageLoading.
