+++ Sources/PackageLoading/PackageBuilder.swift
@@ -230,6 +230,25 @@ public final class PackageBuilder {
         self.createREPLProduct = createREPLProduct
     }
 
+    /// Loads a package from a package repository using the resources associated with a particular `swift` executable.
+    ///
+    /// - Parameters:
+    ///     - package: The absolute path of the package root.
+    ///     - swiftExecutable: The absolute path of a `swift` executable.
+    ///         Its associated resources will be used by the loader.
+    public static func loadPackage(
+        from package: AbsolutePath,
+        with swiftExecutable: AbsolutePath) throws -> Package {
+
+        let manifest = try ManifestLoader.loadManifest(from: package, with: swiftExecutable)
+        let builder = PackageBuilder(
+            manifest: manifest,
+            path: package,
+            diagnostics: DiagnosticsEngine(),
+            isRootPackage: true)
+        return try builder.construct()
+    }
+
     /// Build a new package following the conventions.
     public func construct() throws -> Package {
         let targets = try constructTargets()
